inline remaining code from mongo.integration.base.tests.js
